Reviewed-on: https://codeberg.org/tfkhdyt/blog-api/pulls/7

+++ internal/interface/api/controller/auth_controller.go
@@ -87,7 +87,7 @@ func (a *AuthController) Refresh(c *fiber.Ctx) error {
 		return exception.NewValidationError(err)
 	}
 
-	response, err := a.authUsecase.Refresh(uint(userId), payload)
+	response, err := a.authUsecase.Refresh(userId, payload)
 	if err != nil {
 		return err
 	}
